backend/requirements.txt
# Core dependencies
langchain>=0.0.267
python-dotenv>=1.0.0
redis>=4.5.5
faiss-cpu>=1.7.4
langchain-community>=0.0.1
langchain-huggingface>=0.0.1

# AI models and integrations
google-generativeai>=0.3.1,<0.4.0
langchain-google-genai>=0.0.1,<0.0.2

# PDF handling and text processing
pymupdf>=1.22.3
html2text>=2020.1.16
PyPDF2>=3.0.0
pandas>=2.0.0

# ML/AI models
transformers>=4.41.0
sentence-transformers>=2.6.0
torch>=2.0.0,<3.0.0

# API frameworks
fastapi>=0.68.0
uvicorn>=0.15.0

# Other utilities
tqdm>=4.66.1
numpy>=1.24.3
langdetect
deep-translator
nest-asyncio

# Enhanced PDF extraction
tabulate>=0.9.0